Xdai/Rinkeby support (#64)
As title. Was as easy as updating a few base URLs. Decided to make it an env variable rather than a command line argument, as we likely will use the same value for subsequent queries, thus it seemed to make sense to allow setting it globally for a session.


### Test Plan

```
./gnop trades --count 5 --skip 0
```

Works on default, xdai, rinkeby & when mainnet explicitly set.